feat: move away from process level schannel settings
